fix: http client default state ignores hiddenClients configuration
