+++ packages/api-reference/src/hooks/index.ts
@@ -1,7 +1,6 @@
 export * from './useClipboard'
 export * from './useDarkModeState'
 export * from './useDeprecationWarnings'
-export * from './useHttpClients'
 export * from './useNavState'
 export * from './useOperation'
 export * from './useReactiveSpec'
